refactor keyboard, delete useless KeyboardRu component
