+++ src/bus/client/togglers/index.ts
@@ -9,11 +9,11 @@ import { useSelector } from '../../../tools/hooks';
 // import { Message } from '../../messages/types';
 
 export const initialState = {
-    isOnline:        navigator.onLine,
-    isLoggedIn:      false,
-    isUpperWord:     false,
-    isRussianLayout: false,
-    isKeyboardOpen:  true,
+    isOnline:       navigator.onLine,
+    isLoggedIn:     false,
+    isUpperWord:    false,
+    isRuLayout:     false,
+    isKeyboardOpen: true,
 };
 
 // Types
