+++ src/view/components/Keyboard/styles.ts
@@ -1,33 +1,39 @@
 // Core
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
 // Types
-type booleanProps = {
-    long?: boolean;
-    small?: boolean;
-    smallEnd?: boolean;
-    backspace?: boolean;
+type styleProps = {
+    isRuLayout?: boolean
 }
 
 export const Container = styled.section`
 `;
 
-export const KeyboardWrapper = styled.div`
+export const KeyboardWrapper = styled.div<styleProps>`
   display: grid;
   grid-template-columns: repeat(10, 1fr);
+  ${(props) => props.isRuLayout && css`
+      grid-template-columns: repeat(11, 1fr);
+    `
+}
 `;
 
-export const KeyboardWrapperSmall = styled.div`
+export const KeyboardWrapperSmall = styled.div<styleProps>`
   display: grid;
   grid-template-columns: repeat(9, 1fr);
+
+  ${(props) => props.isRuLayout && css`
+      grid-template-columns: repeat(11, 1fr);
+    `
+}
 `;
 
-export const KeyboardWrapperLast = styled.div`
+export const KeyboardWrapperLast = styled.div<styleProps>`
   display: grid;
   grid-template-columns: repeat(2, 1fr) 50% repeat(2, 1fr);
 `;
 
-export const KeyboardElem = styled.button<booleanProps>`
+export const KeyboardElem = styled.button`
   display: grid;
   place-items: center;
   padding: 3px;
@@ -37,6 +43,7 @@ export const KeyboardElem = styled.button<booleanProps>`
   background-color: #ccc;
   border: 4px solid #000;
   color: #000;
+
     &:hover {
       border-color: #fff;
     }
