+++ src/view/components/index.ts
@@ -8,5 +8,4 @@ export * from './Chat';
 export * from './MessageInfoActions';
 export * from './MessageDateInfo';
 export * from './Keyboard';
-export * from './KeyboardRussian';
 export * from './MainKeyboard';
